dashboard/backend/pyproject.toml
[project]
name = "inside-sales-dashboard-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "google-api-python-client>=2.150.0",
    "google-auth>=2.29.0",
    "google-auth-httplib2>=0.2.0",
    "pydantic>=2.7",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
